Delete hecho

+++ src/main/java/com/salesianostriana/dam/projectFOODAPP/categoria/service/CategoriaService.java
@@ -3,6 +3,7 @@ package com.salesianostriana.dam.projectFOODAPP.categoria.service;
 
 import com.salesianostriana.dam.projectFOODAPP.categoria.dto.GetCategoriaDto;
 import com.salesianostriana.dam.projectFOODAPP.categoria.error.EmptyCategoryWithProductsException;
+import com.salesianostriana.dam.projectFOODAPP.categoria.exception.CategoriaConProductosException;
 import com.salesianostriana.dam.projectFOODAPP.categoria.exception.CategoriaNotFoundException;
 import com.salesianostriana.dam.projectFOODAPP.categoria.exception.EmptyCategoriesException;
 import com.salesianostriana.dam.projectFOODAPP.categoria.model.Categoria;
@@ -70,10 +71,14 @@ public class CategoriaService {
     public void deleteCategoria (String nombreCategoria) {
 
         Categoria cat = categoriaRepository.findByNombreIgnoreCase(nombreCategoria);
+        int cantProductos = categoriaRepository.contarCantidadProductosDeUnaCategoriaByNombre(nombreCategoria);
 
         if (cat == null)
             throw new CategoriaNotFoundException(nombreCategoria);
 
-        categoriaRepository.delete(cat);
+        if (cantProductos == 0)
+            categoriaRepository.delete(cat);
+        else
+            throw new CategoriaConProductosException(nombreCategoria);
     }
 }
